Make a library out of wasi-stub
